+++ pysat/utils/io.py
@@ -475,6 +475,7 @@ def load_netcdf_pandas(fnames, strict_meta=False, file_format='NETCDF4',
                             nc_key)
                     meta[key] = meta_dict
 
+                # TODO(#913): Remove 2D support
                 if len(data.variables[key].dimensions) == 2:
                     # Part of a DataFrame to store within the main DataFrame
                     two_d_keys.append(key)
@@ -485,7 +486,8 @@ def load_netcdf_pandas(fnames, strict_meta=False, file_format='NETCDF4',
                                                'data in pandas. Please use',
                                                'xarray for this file.')))
 
-            # We now have a list of keys that need to go into a DataFrame,
+            # TODO(#913): Remove 2D support
+            # We now have a list of keys that need to go into a dataframe,
             # could be more than one, collect unique dimensions for 2D keys
             for dim in set(two_d_dims):
                 # First or second dimension could be epoch. Use other